Refactoring Product Service #1

+++ src/hohland_cblp.ShopBackend.Domain/Entities/Product.cs
@@ -4,7 +4,7 @@ public record Product()
 {
     public long Id { get; init; }
     public string Name { get; init; }
-    public float Price { get; init; }
+    public decimal Price { get; init; }
     public string Currency { get; init; }
     public ProductType ProductType { get; init; }
     public DateTime CreationDate { get; init; }
